lc-14 - Fixed spring security and added unit tests
